+++ GPU/CAGPUTracking/Standalone/qa/genEvents.cpp
@@ -72,7 +72,7 @@ double genEvents::GetGaus( double sigma )
   double x = 0;
   do{
     x = gRandom->Gaus(0.,sigma);
-    if( fabs(x)<=3.5*sigma ) break;
+    if( fabsf(x)<=3.5*sigma ) break;
   } while(1);
   return x;
 }
@@ -208,7 +208,7 @@ int genEvents::GenerateEvent(const AliGPUCAParam& sliceParam, char* filename)
 	  t.Print();
 	  break;
 	}
-	if( fabs(t.GetZ())>=250. ){
+	if( fabsf(t.GetZ())>=250. ){
 	  std::cout<<"Can not propagate to x = "<<xRow<<": Z outside the volume"<<std::endl;
 	  t.Print();
 	  err = -1;
